# perf: log mergeProps 'wasted' time, total time and number of invocations