When choosing the existing 'Debug' --> 'Toggle react profiling' menu option, also include the following:
- all time that is spent in all component mergeProps functions that results in no properties actually changing (and therefore no render), every 1 second. This is an important metric to improve since this is blocking the renderer window UI thread, and the time is likely spent re-computing functions even though the state input to those functions has not changed.
- total time for each Component-type spent in mergeProps, output every 10 seconds
- number of invocations for each Component-type mergeProps, output every 10 seconds

This measures the issue described by #11515, and allows us to compare the before and after compared to the solution proposed in #12105.

Using these features I found that:
- with about 10 tabs open and switching between them, we're **wasting** about a second of unnecessary blocking UI thread time, every 10 seconds (time taken in mergeProps that resulted in no changes to props).
- when typing in the url bar, we're blocking for a total of about 330ms every 10 seconds, and 90% of that is time wasted (resulted in no changed to component props.

![image](https://user-images.githubusercontent.com/741836/36240276-7d686440-11c5-11e8-97e2-f67977e41460.png)


## Test Plan:
- Ensure that **no** extra logging is showing in the console without choosing the 'Debug' --> 'Toggle react profiling' and also when toggling it on and off again.
- Ensure that the extra logging **is** showing in the console without choosing the 'Debug' --> 'Toggle react profiling' and also when toggling it off and on again.

## Submitter Checklist:

- [ ] Submitted a [ticket](https://github.com/brave/browser-laptop/issues) for my issue if one did not already exist.
- [ ] Used Github [auto-closing keywords](https://help.github.com/articles/closing-issues-via-commit-messages/) in the commit message.
- [ ] Added/updated tests for this change (for new code or code which already has tests).
- [x] Ran `git rebase -i` to squash commits (if needed).
- [x] Tagged reviewers and labelled the pull request [as needed](https://github.com/brave/browser-laptop/wiki/Pull-request-process).
- [ ] Request a security/privacy review [as needed](https://github.com/brave/handbook/blob/master/development/security.md#how-to-request-a-security-review). (Ask a Brave employee to help if you cannot access this document.)

## Reviewer Checklist:

- [ ] Request a security/privacy review [as needed](https://github.com/brave/handbook/blob/master/development/security.md#how-to-request-a-security-review) if one was not already requested.

Tests


- [ ] Adequate test coverage exists to prevent regressions
- [ ] Tests should be independent and work correctly when run individually or as a suite [ref](https://github.com/brave/browser-laptop/wiki/Code-Guidelines#test-dependencies)
- [ ] New files have MPL2 license header