@petemill 2018-02-16
Addressed @NejcZdovc's feedback by making sure nothing is logged if there is no time spent in mergeProps at all within the (10 second) period. This makes sure that if you step out for lunch, you don't come back to hundreds of '0' log entries. 🍏